Checking whether the URL should be manually changed after a transition
RC7 includes a check which prevents the URL from being updated when transitioning to identical routes - however updates may be needed when the query string has changed. Fixes #6.

NB: There is a cleaner way of solving this by checking for changes inside `updateParameterContexts`, but that would break compatibility with <RC7.